requirements.txt
# ── Telegram бот ──────────────────────────────────────────────────────────────
aiogram==3.4.1

# ── Google AI (Gemini) ────────────────────────────────────────────────────────
google-generativeai>=0.8.3

# ── Firebase ──────────────────────────────────────────────────────────────────
firebase-admin==6.4.0

# ── Парсинг HTML ──────────────────────────────────────────────────────────────
beautifulsoup4==4.12.3
lxml==5.2.1          # быстрый HTML-парсер (резерв для bs4)
requests==2.31.0

# ── RAG / Векторный поиск ─────────────────────────────────────────────────────
numpy>=1.26.0
# langchain и langchain-openai удалены для перехода на нативный SDK Google

# ── Планировщик задач ─────────────────────────────────────────────────────────
apscheduler==3.10.4

# ── Валидация / конфигурация ──────────────────────────────────────────────────
pydantic>=2.4.1,<2.6
python-dotenv==1.0.1
